refactor: update feature components tests
